server/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
configparser==3.5.0
Flask==1.0.2
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
newsapi-python==0.2.3
python-dateutil==2.7.5
requests==2.17.1
six==1.11.0
urllib3==1.21.1
watson-developer-cloud==2.4.1
websocket-client==0.47.0
Werkzeug==0.14.1
